Performing Batch Updates - P1
- We want to create a method where we can actually instead of having to
maybe call our "create" or our "addFoodGroup" methods multiple times in
a loop, if we want
to do something like add multiple records, we are going to create a
method that can actually handle this using "batchUpdate" function that
provided by Spring.

- In this first part: we would write the whole java code of how we
extract batch parameters from a list of objects (list of foodGroup).
However, in the second part, we will just use one Spring instruction to
do so.